modules/locations-api/requirements.txt
grpcio==1.42.0
grpcio-tools==1.42.0
kafka-python==2.0.2
psycopg2-binary==2.9.2